MDL-58768 calendar: Add $beneficiaryuser to the container class
The beneficiary user is not always the same as the current user. The container class now allows a different user
to be set as the beneficiary user. All capability checks are done against the beneficiary user.

Also, the $beneficiaryuser is passed to core_calendar_provide_event_action and core_calendar_is_event_visible
callback functions. These callback functions need to be updated in all activity modules to accept a $user parameter.