MDL-63809 question: Proper text for broken questions during an attempt
It was a jumping conclusion bias to assume that the question is surely
deleted whenever there is an exception during loading a question.
Changed the text to indicate that the question may also be broken
(rather than saying it must have been deleted). Also, a debugging
message is being displayed when we encounter a broken question.