MDL-71603 editor_atto: Focus on the toggled row
quoted from
https://www.accessibility-developer-guide.com/examples/sensible-aria-usage/expanded/

As a general rule, the toggled element should be right below the toggle
button, so screen readers will find it easily. If that is not the case,
then the focus should be placed inside the element upon toggling it
visible, and back to the initial element upon toggling it invisible.